[metadata] consolidate MetadataWrapper methods to the same file
